Refactored duplicates between TOE, FRY and MEB with the shared directory
