+++ src/applications/my-education-benefits-apps/my-education-benefits/utils/validation.js
@@ -1,6 +1,6 @@
 // import { isValidEmail } from 'platform/forms/validations';
-import moment from 'moment';
-import { formatReadableDate } from '../helpers';
+// import moment from 'moment';
+// import { formatReadableDate } from '../helpers';
 import { validatePhone } from '../../shared/validations';
 
 // export const isValidPhone = (phone, isInternational) => {
@@ -40,19 +40,19 @@ export const validateMobilePhone = (errors, phone, formData) => {
 //   }
 // };
 
-export const validateEffectiveDate = (errors, dateString) => {
-  const effectiveDate = moment(dateString);
-  const minDate = moment().subtract(1, 'year');
-  const maxDate = moment().add(180, 'day');
-
-  if (
-    effectiveDate.isBefore(minDate, 'day') ||
-    effectiveDate.isAfter(maxDate, 'day')
-  ) {
-    errors.addError(
-      `Please enter a date between ${formatReadableDate(
-        minDate.format('YYYY-MM-DD'),
-      )} and ${formatReadableDate(maxDate.format('YYYY-MM-DD'))}`,
-    );
-  }
-};
+// export const validateEffectiveDate = (errors, dateString) => {
+//   const effectiveDate = moment(dateString);
+//   const minDate = moment().subtract(1, 'year');
+//   const maxDate = moment().add(180, 'day');
+//
+//   if (
+//     effectiveDate.isBefore(minDate, 'day') ||
+//     effectiveDate.isAfter(maxDate, 'day')
+//   ) {
+//     errors.addError(
+//       `Please enter a date between ${formatReadableDate(
+//         minDate.format('YYYY-MM-DD'),
+//       )} and ${formatReadableDate(maxDate.format('YYYY-MM-DD'))}`,
+//     );
+//   }
+// };
